Extract MCP work-proof guidance helpers
